Added heroes screen and updated service to use the user defined bounds.
The heroes screen is to help users and devs debug why a particular hero may not be processable- i.e. isn't present in data, name issue, etc.

The bounding config bug was related to the logic in the service class. The service class always ended up using the default config instead of the user defined one. This has been fixed.