Fixes to stream handling and errors.
Since I was using doOnError and doOnSubscribe, I was not actually assigning handlers for success/errors. As a result, errors were registered as uncaught. The proper approach is to provide an explicit onError and onSuccess callback in the subscribe() call.

Additionally, this fixes some uncaught error cases, which led me to implement a toast for the user on hero details if an error occurs.

Also, made some other minor improvements:
- Calling Activity.recreate instead of sending an intent to the same activity
- updated proguard rules for Glide
- updated build.gradle